bmp180: Add fallback EEPROM values, and increase i2c retries.
I have observed that the EEPROM values sometimes cannot be read. I'm not sure
exactly what's wrong, but being unable to read the EEPROM doesn't appear to go
hand-in-hand with not being able to read the pressure and temperature.

These EEPROM values are dumped from my particular BMP180. I want to get some
more EEPROM dumps to get a range of typical values so we can find a
middle-ground to at least get acceptable data on as many sensors as we can.

I have a feeling the problem may be related to the fact that i2c reads in
this driver try to read a full block of data. The EEPROM read has a large
length compared to the other typical i2c reads. This needs further
investigation still.

Signed-off-by: Kyle Repinski <repinski23@gmail.com>